+++ lib/s_extensions.dart
@@ -3,3 +3,4 @@ library s_extensions;
 export 'extensions/number_ext.dart';
 export 'extensions_context.dart';
 export 'extensions/string_ext.dart';
+export 'extensions/datetime_ext.dart';